Cleanup and added additional exception class

+++ src/main/java/com/lukka/notifybackend/service/UserService.java
@@ -1,5 +1,6 @@
 package com.lukka.notifybackend.service;
 
+import com.lukka.notifybackend.exception.EmptyRepositoryException;
 import com.lukka.notifybackend.exception.ResourceNotFoundException;
 import com.lukka.notifybackend.model.User;
 import com.lukka.notifybackend.repo.UserRepo;
@@ -16,27 +17,38 @@ public class UserService {
     private UserRepo userRepo;
 
     // POST
-    public User save(User user){
+    public User save(User user) {
         return userRepo.save(user);
     }
+
     // GET
     public List<User> getAllUsers() {
-        return userRepo.findAll();
+        List<User> users = userRepo.findAll();
+        if (!users.isEmpty())
+            return users;
+        else
+            throw new EmptyRepositoryException("GET", "User");
     }
 
     public User getUser(String email) {
         return userRepo.findById(email).orElseThrow(() ->
                 new ResourceNotFoundException("User", "email", email));
     }
+
     // DELETE
     public void deleteUser(String email) {
         userRepo.findById(email).orElseThrow(() ->
                 new ResourceNotFoundException("User", "email", email));
         userRepo.deleteById(email);
     }
+
     public void deleteAllUsers() {
-        userRepo.deleteAll();
+        List<User> users = userRepo.findAll();
+        if (!users.isEmpty())
+            userRepo.deleteAll();
+        else throw new EmptyRepositoryException("DELETE", "User");
     }
+
     // PUT
     public User updateUser(User user, String email) {
         User existingUser = userRepo.findById(email).orElseThrow(() ->
